+++ ui/mods/commander_image_color/start.js
@@ -1,9 +1,6 @@
 (function() {
   var cic = commander_image_color
 
-  var baseBlue = [58, 119, 174]
-  var baseYellow = [255, 200, 2]
-
   var colorize = function() {
     var primaryChoice = cic.colorNames[Math.floor(Math.random() * 11)]
     //var primaryChoice = 'BLACK'
@@ -24,7 +21,6 @@
   }
 
   var $commander = $('#commander-img')
-  var offset = $commander.offset()
   var $clone = $commander.clone()
     .attr('id', 'background_commander_clone')
     .attr('data-bind', '')